Stopped prompting users to review app upon completion of event.
Appirater has the option of prompting a user to rate the app once the
app meets all the defined criteria. Since one of our criteria is a
certain amount of significant events this resulted in the app
prompting users as soon as a post was published, or a comment was
liked, etc. and this was a bit of a jarring experience. Defaulting to
asking the user upon the app opening after meeting all the criteria.