Adjusting push notification registration to work with Helpshift.
Previously we would register for push notifications upon sign in of
an account. The problem is that users may want to use Helpshift to
contact us before they are signed in(especially if they are having
sign in problems) and without push notifications enabled they won't be
notified when we respond to them. I adjusted the way we register push
notifications to account for this.